+++ receive.py
@@ -10,12 +10,40 @@ Currently, all this file does is check for received emails over the last 24
 hours from [email] and with the words "diary email" in the Subject. It stores
 the contents of these emails in S3 container.
 """
+import json
 import os
+from typing import TypedDict
+from urllib.parse import urlencode
+from urllib.request import urlopen
+
+from dotenv import load_dotenv
 from imap_tools import MailBox
-from main import source_env
+
+
+class Token(TypedDict):
+    access_token: str
+    token_type: str
+    expires_in: int
+
+
+def refresh_access_token(client_id, client_secret, refresh_token) -> Token:
+    """Generate a refreshed access token.
+
+    Inspired by:
+        https://github.com/kootenpv/yagmail/blob/f24af871c670c29f30c34ef2a4ab5abc3b17d005/yagmail/oauth2.py#L63
+    Background information:
+        https://developers.google.com/youtube/v3/live/guides/auth/server-side-web-apps#OAuth2_Refreshing_a_Token
+    """
+    params = dict(client_id=client_id, client_secret=client_secret,
+                  refresh_token=refresh_token, grant_type="refresh_token")
+    encoded_params = urlencode(params).encode("utf-8")
+    response = urlopen("https://accounts.google.com/o/oauth2/token",
+                       encoded_params).read().decode("utf-8")
+    return json.loads(response)
+
 
 if __name__ == "__main__":
-    source_env()
+    load_dotenv()
 
     sender = os.environ["SENDER"]
     receiver = os.environ["RECEIVER"]
