+++ requirements.txt
@@ -1,2 +1,3 @@
 imap-tools==0.40.0
+python-dotenv==0.17.0
 yagmail==0.14.247
